# add a new AMD gpu name to tiny_ocl.h