AMD Radeon 6700 XT returns "gfx1031" for CL_DEVICE_NAME.

I also tried CL_DEVICE_VENDOR and it returned "Advanced Micro Devices, Inc.", maybe this could be easier to check.