@jbikker 2024-11-20
Perhaps the CL_DEVICE_VENDOR string can be added as a safety net? Unless that is reported for all models.. Sadly I don't have a lot of devices to test on, so I'm grateful for your PR!