@bekorn 2024-11-20
Before fixing this, ISAMD was not defined on traverse.cl > SIMD_AABBTEST was not defined > AABB test used _native_fma > OCL compiler gave the error: "use of undeclared identifier '_native_fma'" > failed speedtest. Which seemed like something that shouldn't happen however, I have no experience with OCL unfortunately, just reporting