@jbikker 2024-11-20
So with your patch properly compiles on AMD?
Question: could you try out which combination of 'SIMD_AABBTEST' and 'USE_VLOAD_VSTORE' is best for AMD? On Intel, performance is best when I don't use the vload/vstore commands but do use the 'SIMD' aabb test code; on NVIDIA it's the other way round. For AMD I am just guessing.