@bekorn 2024-11-20
If I don't use SIMD_AABBTEST, it doesn't compile because `error: "use of undeclared identifier '_native_fma'"`. Full log:
```
C:\Users\Berk\AppData\Local\Temp\comgr-56503e\input\CompileSource:385:22: error: use of undeclared identifier '_native_fma'
  385 |                                                 float tminx0 = _native_fma( lox4.x, idirx, origx ), tminx1 = _native_fma( lox4.y, idirx, origx );
      |                                                                ^
C:\Users\Berk\AppData\Local\Temp\comgr-56503e\input\CompileSource:385:68: error: use of undeclared identifier '_native_fma'
  385 |                                                 float tminx0 = _native_fma( lox4.x, idirx, origx ), tminx1 = _native_fma( lox4.y, idirx, origx );
      |                                                                                                              ^
C:\Users\Berk\AppData\Local\Temp\comgr-56503e\input\CompileSource:386:22: error: use of undeclared identifier '_native_fma'
  386 |                                                 float tminy0 = _native_fma( loy4.x, idiry, origy ), tminy1 = _native_fma( loy4.y, idiry, origy );
      |                                                                ^
C:\Users\Berk\AppData\Local\Temp\comgr-56503e\input\CompileSource:386:68: error: use of undeclared identifier '_native_fma'
  386 |                                                 float tminy0 = _native_fma( loy4.x, idiry, origy ), tminy1 = _native_fma( loy4.y, idiry, origy );
      |                                                                                                              ^
C:\Users\Berk\AppData\Local\Temp\comgr-56503e\input\CompileSource:387:22: error: use of undeclared identifier '_native_fma'
  387 |                                                 float tminz0 = _native_fma( loz4.x, idirz, origz ), tminz1 = _native_fma( loz4.y, idirz, origz );
      |                                                                ^
C:\Users\Berk\AppData\Local\Temp\comgr-56503e\input\CompileSource:387:68: error: use of undeclared identifier '_native_fma'
  387 |                                                 float tminz0 = _nativ
```

With USE_VLOAD_VSTORE:
```
tiny_bvh version 0.9.5 performance statistics (MSVC 1941 build)
running on AMD Ryzen 5 7600 6-Core Processor
----------------------------------------------------------------
available OpenCL platforms:
#0: AMD Accelerated Parallel Processing
OpenCL device: AMD Accelerated Parallel Processing
Device # 0, (name: gfx1031) (platform: OpenCL 2.0 AMD-APP (3617.0)) (vendor: Advanced Micro Devices, Inc.)
Local memory size: 64KB
hardware detected: AMD.
----------------------------------------------------------------
Loading triangle data (262267 tris).
BVH construction speed
warming caches...
BVH traversal speed
- GPU, coherent,   alt 2-way layout,  ocl:      6.7ms for   7.68M rays => 1139.37MRay/s
- GPU, coherent,   alt 4-way layout,  ocl:      6.2ms for   7.68M rays => 1231.07MRay/s
- GPU, coherent,   CWBVH layout,      ocl:      7.3ms for   7.68M rays => 1047.86MRay/s
all done.
```
Without USE_VLOAD_VSTORE:
```
- GPU, coherent,   alt 2-way layout,  ocl:      6.8ms for   7.68M rays => 1133.41MRay/s
- GPU, coherent,   alt 4-way layout,  ocl:      6.2ms for   7.68M rays => 1231.23MRay/s
- GPU, coherent,   CWBVH layout,      ocl:      7.8ms for   7.68M rays => 979.40MRay/s
```