@bekorn 2024-11-20
With your new commits:

```
#define USE_VLOAD_VSTORE
//#define SIMD_AABBTEST
- GPU, coherent,   CWBVH layout,      ocl:      7.3ms for   7.68M rays => 1047.28MRay/s
```

```
//#define USE_VLOAD_VSTORE
//#define SIMD_AABBTEST
- GPU, coherent,   CWBVH layout,      ocl:      7.9ms for   7.68M rays => 976.71MRay/s
```