@bekorn 2024-11-20
I thnik there is a typo here

```
inline float _native_fma( const float a, const float b, const float c )
{
#ifdef _ISNVIDIA <= should be ISNVIDIA
```