@bekorn 2024-11-20
Yes, seems like it. Radeon 6700 XT is equivalent to 4060 Ti, fyi

> Thank you for testing that. So the optimal setting is vload on, simd off for AMD, yielding about 1Gray/s.
> […](#)
> On Wed, Nov 20, 2024 at 6:01 PM bekorn ***@***.***> wrote: With your new commits: #define USE_VLOAD_VSTORE //#define SIMD_AABBTEST - GPU, coherent, CWBVH layout, ocl: 7.3ms for 7.68M rays => 1047.28MRay/s //#define USE_VLOAD_VSTORE //#define SIMD_AABBTEST - GPU, coherent, CWBVH layout, ocl: 7.9ms for 7.68M rays => 976.71MRay/s — Reply to this email directly, view it on GitHub <[#25 (comment)](https://github.com/jbikker/tinybvh/pull/25#issuecomment-2489120743)>, or unsubscribe <https://github.com/notifications/unsubscribe-auth/AFH3PEHM5XZYW73PIGSGCU32BS55BAVCNFSM6AAAAABSEWRMVWVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZDIOBZGEZDANZUGM> . You are receiving this because you modified the open/close state.Message ID: ***@***.***>