@jbikker 2024-11-20
Indeed, thanks.

On Wed, Nov 20, 2024 at 6:04 PM bekorn ***@***.***> wrote:

> I thnik there is a typo here
>
> inline float _native_fma( const float a, const float b, const float c )
> {
> #ifdef _ISNVIDIA <= should be ISNVIDIA
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/jbikker/tinybvh/pull/25#issuecomment-2489128774>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AFH3PEBNSXMK7C6JL6CGSPT2BS6K5AVCNFSM6AAAAABSEWRMVWVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZDIOBZGEZDQNZXGQ>
> .
> You are receiving this because you modified the open/close state.Message
> ID: ***@***.***>
>