@bekorn 2025-01-04
I stumbled upon this dropdown in RenderDoc, not sure if it is exhaustive though.
![image](https://github.com/user-attachments/assets/8cc92836-07c0-49f2-9e2c-7108763f6d69)